+++ NEWS.md
@@ -4,6 +4,7 @@ increments reflect backwards-incompatible API changes, not necessarily big chang
 
 # huxtable 4.3.0.9000
 
+* New `set_outer_border_colors()` and `set_outer_border_styles()` functions, like `set_outer_borders()`.
 * Bugfix: `tidy.tidy_override()` and `glance.tidy_override()` should work even if underlying object
   has no `tidy()` or `glance()` method.
 * Bugfix: huxtables had option clash when `echo = TRUE` in Rmd pdf_document format.